[POLARIS] BE Fix : Trade id 자료형 Long으로 변경

+++ back-end/polaris/src/main/java/com/ssafy/polaris/chat/repository/TradeRepository.java
@@ -12,7 +12,7 @@ import org.springframework.data.repository.query.Param;
 
 import jakarta.transaction.Transactional;
 
-public interface TradeRepository extends JpaRepository<Trade, Integer> {
+public interface TradeRepository extends JpaRepository<Trade, Long> {
 	// 나의 채팅방 목록
 	@Query(value =
 		"SELECT " +
